fix name on link and add castom hook to hooks
