+++ src/components/BookingForm/BookingForm.module.scss
@@ -1,7 +1,6 @@
 @import '../../App.styles';
 
 .booking {
-  //max-width: 750px;
   width: 100%;
   display: flex;
   border-radius: 15px;
