+++ src/routes/DashboardRoutes.tsx
@@ -27,8 +27,7 @@ const DashboardRoutes: FC = () => {
                 <Route path="timeline" element={<Timeline/>}/>
                 <Route path="calendar" element={<Calendar/>}/>
                 <Route path="settings" element={<Settings/>}/>
-                <Route path="bookig_form_page" element={<BookingFormPage/>}/>
-
+                <Route path="bookig-form" element={<BookingFormPage/>}/>
             </Route>
         </Routes>
     );
